docs: Balder and Bruggen's paper
* Summary: This information paper presents the work being undertaken by the
CANSO Safety Performance Measurement Work Group in the development of
time-based risk detection tools. The tools are designed to complement the
Risk Analysis Tool methodology and to provide a more complete picture of an
Air Navigation Service Provider’s mid-air collision risk and a method for the
identification and monitoring of key airspace risk areas.